added authorization by email
