+++ src/components/CategoryList.tsx
@@ -12,6 +12,7 @@ import RemoveCategoryModal from './modals/RemoveCategoryModal';
 import EditCategoryModal from './modals/EditCategoryModal';
 import {ToastError, ToastSuccess} from '../utilities/error-handling';
 import CreateCategoryModal from './modals/CreateCategory';
+import {isAdmin} from "../constants/isAdmin";
 
 export const CategoryList = () => {
     const [isLoading, setIsLoading] = useState(false);
@@ -21,8 +22,6 @@ export const CategoryList = () => {
     const {currentCategory, categories, onChangeCurrentCategory, onChangeCategories} = useCategory();
     const [selectedCategory, setSelectedCategory] = useState({} as ICategory);
     const [error, setError] = useState('');
-
-    const isAdmin = false;
 
     const fetchCategories = async () => {
         setIsLoading(true)
