+++ src/constants/isAdmin.ts
@@ -1,0 +1,1 @@
+export const isAdmin = true;
