+++ src/models/IUser.tsx
@@ -1,0 +1,4 @@
+export interface IUser {
+    email: string,
+    password: string
+}
